requirements.txt
annotated-types==0.7.0
attrs==24.2.0
backports-datetime-fromisoformat==2.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
construct==2.10.70
ffmpeg-python==0.2.0
future==1.0.0
h11==0.14.0
idna==3.10
m3u8==6.0.0
numpy==2.1.3
opencv-python==4.10.0.84
outcome==1.3.0.post0
pillow==11.0.0
pydantic==2.10.2
pydantic_core==2.27.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
matplotlib
transformers
scipy
llama-index-core
llama-index-llms-openai
fastapi
uvicorn
python-multipart
llama-index-agent-openai